old/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antspyx==0.5.4
appnope==0.1.4
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
datasets==3.2.0
debugpy==1.8.6
decorator==5.1.1
Deprecated==1.2.15
dill==0.3.8
dm-tree==0.1.8
einops==0.8.1
exceptiongroup==1.2.2
execnet==2.1.1
executing==2.1.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
fractalbrain==1.0
frozenlist==1.5.0
fsspec==2024.9.0
gast==0.6.0
gdown==4.7.1
glob2==0.7
google-pasta==0.2.0
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
imageio==2.36.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.0.0
IProgress==0.4
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kagglehub==0.3.6
keras==3.6.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.43.0
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.3.2
monai-weekly==1.5.dev2448
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
namex==0.0.8
natsort==8.4.0
nest_asyncio==1.6.0
networkx==3.4.2
nibabel==3.2.2
nilearn==0.11.1
numba==0.60.0
numpy==1.26.0
opencv-python==4.8.1.78
opt_einsum==3.4.0
optree==0.13.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.1
pip==24.3.1
pip-system-certs==4.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==18.0.0
pycad-medic==0.0.10
pydicom==2.4.3
Pygments==2.18.0
pynrrd==1.1.1
pyparsing==3.1.4
PySocks==1.7.1
pytest==8.3.3
pytest-shutil==1.8.0
python-dateutil==2.9.0
pytorch-ignite==0.4.11
pytz==2024.2
pyvista==0.44.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.3
scikit-image==0.22.0
scikit-learn==1.5.2
scipy==1.14.1
scooby==0.10.0
seaborn==0.13.2
setuptools==75.1.0
shap==0.46.0
SimpleITK==2.3.0
six==1.16.0
slicer==0.0.8
soupsieve==2.6
stack-data==0.6.2
statsmodels==0.14.4
sympy==1.13.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-probability==0.25.0
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.2.2
tornado==6.4.1
tqdm==4.66.1
traitlets==5.14.3
trimesh==4.5.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vedo==2023.4.6
vtk==9.2.6
wcwidth==0.2.13
webcolors==24.11.1
Werkzeug==3.0.6
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xgboost==2.1.3
xxhash==3.5.0
yarl==1.17.0
zipp==3.20.2
